+++ src/stats_duck_extension.cpp
@@ -15,6 +15,7 @@
 #include "bootstrap_function.hpp"
 #include "auto_bin_function.hpp"
 #include "corr_matrix_function.hpp"
+#include "lm_function.hpp"
 #include "table_one_function.hpp"
 #include "normality_function.hpp"
 #include "ks_test_function.hpp"
@@ -72,6 +73,9 @@ static void LoadInternal(ExtensionLoader &loader) {
 
 	// Pairwise correlation matrix
 	RegisterCorrMatrix(loader);
+
+	// OLS linear regression
+	RegisterLm(loader);
 
 	// Table 1 helper
 	RegisterTableOne(loader);